requirements.txt
numpy
pillow
torch
matplotlib
scipy
appdirs
scikit-learn
umap-learn
